Add files via upload
RGB calculation and bar plot verification protocol aims to ensure the efficiency and accuracy of RGB calculation and generation of bar plots for the test strips. This EVP outlines the validation process that is used for the normalization technique applied to RGB values obtained from test strip images. The EVP goes into the detailed steps involved in normalization, and the hypothesis to be tested. Further materials/software are required for the procedure for normalization and data analysis, experimental justification, and the limitations. Two test cases are been shown for comparison of normalized RGB values.