Add files via upload
This EVP outlines the validation process that is used for evaluating the effectiveness of the K-means clustering technique in image processing. The protocol involves loading the image on the image processing environment and then applying K-means clustering to the pixel data. Then comparing and visualizing the original uploaded image and the clustered image which is segmented and grouped into different colors to catch the nuances of the variations.